requirements.txt
asgiref==3.7.2
botocore==1.34.11
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.3.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
packaging==23.2
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
whitenoise==6.6.0
